feat(HU2018): Add ErrorComponent for displaying error messages
Added a new component called ErrorComponent to handle the display of error messages. This component renders an error icon, a message, and a button to reload the page. It is designed to be used in cases where something goes wrong in the application. The ErrorComponent is imported and used in other parts of the codebase to handle error states.